# trees: speed-up training