@Sentimentron 2018-05-10
Tests failed because GoNum seem to have dropped support for Go 1.7, LGTM